댓글 저장 후 리다이렉트 완료, 댓글 수정 삭제 필요

+++ furry_friend/src/main/java/com/ff/furry_friend/repository/CommentRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Repository
 public interface CommentRepository {
     //댓글
-    List<comment> findComment(int pro_id);
+    List<comment> findComment(Long pro_id);
 
     //댓글 저장
     void save(comment comment);
