+++ furry_friend/src/main/java/com/ff/furry_friend/service/CommentService.java
@@ -17,7 +17,7 @@ public class CommentService {
         this.commentRepository = commentRepository;
     }
 
-    public List<comment> findComment(int pro_id){
+    public List<comment> findComment(Long pro_id){
         return commentRepository.findComment(pro_id);
     }
